moved savedataeditor to own file
